requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
requests
bcrypt
fpdf2
pandas-ta
matplotlib.pyplot
matplotlib 